Pick ups
I have done the 3 pick up code for the game but, it still needs particle
effects and the images to be resized or redone so that it fits properly
and need different colours. I will work on the particle effects unless I
can't seem to be able to do it.