Added Particle Effects
I have added particle effects to the pick up and made the lotus images
smaller in paint (because I don't know how to make it smaller in Unity
for now). Also I have edited the lotus code a bit so that it gives the
correct amount of chi or health back to the player.